+++ src/main/scala/mesosphere/marathon/GpuSchedulingBehavior.scala
@@ -6,12 +6,6 @@ sealed trait GpuSchedulingBehavior {
 object GpuSchedulingBehavior {
   case object Undefined extends GpuSchedulingBehavior {
     override def name = "undefined"
-    if (BuildInfo.version < SemVer(1, 9, 0))
-      "undefined"
-    else {
-      /* See MARATHON-8589 */
-      throw new NotImplementedError("Hi programmer! Please delete the definition Undefined and fix all resulting compile issues.")
-    }
   }
 
   case object Restricted extends GpuSchedulingBehavior {